perfetto-ui: Fix js stack overflow
Fix a couple of instances where we used:

someArray.push(...someOtherArray);

to append all elements from someOtherArray to someArray as
this (I now realaise) is an anti-pattern. If someOtherArray is very
large this code will cause a stackoverflow attempting to pass all the
elements as arguments.

Also fix an N+1 query pattern in trace_controller that was slow when
there are many tracks.

Bug: 144595839
Change-Id: I5fa8773af66adb617578795c68da8ccfe1035504