trace_processor: rewrite span join to be clearer
This CL rewrites span join to be a lot easier to read and understand
by encoding the queries explicitly as a state machine and always emitting
shadow slices. By doing this and then post filtering, we reduce the number
of edge cases

By doing this, we also fix a subtle issue I spotted with left join not
correctly outputting slices where there is no partition in the other table.

Change-Id: Ie89c5717bf80d5b73664fba810586a33da2643ff