trace_processor: fix null dereference in ref_type mapping
As we were calling InternString with a nullptr, we were getting crashes
in strlen when looking up the string for a reftype.

Instead of storing char*, store the StringView which lets us avoid the
issue - this code is going away soon anyway so trying to fix this better
is not worth our time.

This fixes the Chrome roll.

Change-Id: I1005b64128fee772a80b0f0a2bb68f67d3dce791