Replace lite-server with serveit clone
Replace lite-server (which has 287 dependencies http://npm.broofa.com/?q=lite-server)
with a 134 line Python script.

The replacement is not quite like for lite as serveit does not continuously watch files for
changes. Rather whenever you refresh the page from the server it anything has changed it
checks to see if any file has been changed, if it has is blocks the request until the
rebuild command is done.

Change-Id: I29f00f7787a8e8c700665f1af0220d577c0aa1ee