perfetto-ui: Show CPU slice aggregation
Hold shift and do an area selection across some tracks.
In the details panel you can see the aggregation of the selected
slices.

Also fixed the details panel height bug. We don't want to remember
different heights for each type of details panel info because that
will feel inconsistent and confusing for the user. Instead we just
need an easy way to make the details panel full screen. This CL adds
a button for that.

Bug:118899922
Change-Id: Ic799411f4b6706d65bbc4c9300d5031d340c5c57