trace_processor: add a fast path for zero/one row filtered tables
This CL adds a fast path inside the db sqlite glue code in cases where
the filtered RowMap only has a single row.

Instead of applying this RowMap to the source table and then creating
an iterator on the table, we can just lookup the single row directly and
use that to directly lookup inside the source table.

By doing this, we remove the allocations of columns and rowmaps inside
the table as well as the iterators inside table's iterator.

This is necessary to remove the regression observed when migrating the
thread and process tables.

Context: go/perfetto-tp-refactor
Bug: 135177627
Change-Id: Ia3363898a04e9863aab5f339b0a05d9a21503962