trace_processor: add parsing and ingestion of mm_id of rss_stat
This CL adds support for parsing (in traced_probes) and ingesting mm_id
field of rss_stat which lets us track cross-process rss changes (that is
changes of an mm struct of process A by another process B).

In the past, these sort of changes were inaccurately tracked. With mm_id
and is_curr, we can accurately attribute it to the correct process and
drop events when we don't know the process whose struct is being
changed.

We also add tests to check both the old and new behaviour.

Bug: 144825618
Change-Id: If9ce3a2215e364335130c40feb923009f6a660eb