proto: Fix field ID inconsistency in TrackEvent/TrackEventDefaults
Field IDs in the defaults should be synchronized with their
corresponding regular message. In the case of track_uuid, this wasn't
the case.

Since noone should be relying on the defaults yet (Chrome doesn't emit
them yet, and client lib always emits the non-default value too), it
should be safe to simply fix the field ID.

Change-Id: Icac9038b40e4654029dd39cb35a8f0b47080313a