heapprofd producer: use sigqueue(si_value: 0) instead of kill
We want to multiplex the signal (as traced_perf has similar needs), and
so have to use si_value to discriminate between the two.

As far as I can tell, si_value is usually uninitialized when receiving a signal
sent through kill, so the zero value is not strictly for backwards compat, but
rather just the first value we reserve.

Bug: 144281346
Change-Id: I0a60c835458b00cf357ec1160336ce6122660ae6