processor: Refactor string pool IDs
We recently added a way to specify a smaller block size on 64bit systems
for Chrome. However, we overlooked that StringIds on 64bit don't support
more than one block in their current state.

This refactors StringPool and its IDs to store a block index and offset
inside an Id and translate it back into those when looking up. Large
strings are stored in a separate vector and store the index into this
vector in their Ids. The two different kinds of Ids are distinguished
using the MSB of the Id.

Bug: chromium:1041573
Change-Id: Ied70cb06febdcbdb86008df4d5f18b69e52ae2fd