ui extension: Allow connections from localhost
When we added the extension originally, it seemed like we needed to
proxy "localhost" via another hostname/domain to allow a local UI build
to connect to the extension. It seems this is actuall not necessary
(anymore?), based on local testing on canary. (I think we might have
attempted to specify a port in the url filter last time, but that's not
necessary.)

Let's try allowing localhost connections directly instead.

Change-Id: Ib70ac0500b4a8a340d5f045baf9b11d569fa9b67