GTEST_EXECUTE_STATEMENT_ is not available on iOS execute the statement directly instead.
GTEST_EXECUTE_STATEMENT_ is used for GTest Death Test, and is only
available on systems that support it. So if it is available use it for the same
behaviour today.

GTest says[1] the following about GTEST_EXECUTE_STATEMENT_:

"
// This macro is for implementing ASSERT/EXPECT_DEBUG_DEATH when compiled in
// NDEBUG mode. In this case we need the statements to be executed and the macro
// must accept a streamed message even though the message is never printed.
// The regex object is not evaluated, but it is used to prevent "unused"
// warnings and to avoid an expression that doesn't compile in debug mode.
"

For us since we aren't in DEBUG (we would use the IF_SUPPORTED macro instead) these
statements should not die in DEBUG we can just execute the statement directly
and they might be optimized away on iOS but this isn't a huge concern if one
platform has less coverage.

[1] https://github.com/google/googletest/blob/master/googletest/include/gtest/internal/gtest-death-test-internal.h

Bug: 147743969
Change-Id: I6097aec83391f13101376afe5261104cc281b370