GN: clean up build files, split ipc client/service
This CL refactors the build files to allow a cleaner
separation of producer, consumer and service code by
later CLs. The end goal is to allow clients to link
a reduced version of the client library with only the
the producer code in Android (i.e. no kInProcess mode)
to reduce binary size and memory usage (see bug).
This CL alone doesn't reduce the binary size but just
puts the GN files in a better state. Changes introduced:

1. src/tracing: create GN files for each sub-directory
   instead of having one monster BUILD.gn.

2. Split the src/ipc targets into client vs host.

3. Split the src/tracing/ipc targets into
   producer, consumer, service.

4. Some further minor cleanups.

Bug: 148198993
Change-Id: Ie501c930255b10b8eafcd61b7226cd49e32728be