tp: fix bad performance when iterating on small out range
This CL fixes the degenerate performance when we have a small output
range with very large indices - this would lead to a giant bitvector
being generated which would domniate performance.

Change this so that either when we know that an index vector must be
cheaper or when the range is quite small to just use the indexvector.

Before:
                                               Time         Iterations
 -------------------------------------------- ------- ---- ------------
  BM_TableFilterParentSortedAndOther/2097152   11063   ns        60440
  BM_RowMapFilterIntoRangeWithRange            26333   ns        26688

After:
                                               Time        Iterations
 -------------------------------------------- ------ ---- ------------
  BM_TableFilterParentSortedAndOther/2097152    798   ns       884238
  BM_RowMapFilterIntoRangeWithRange            2282   ns       303897

Bug: 148135145
Change-Id: I94b121682786ae10b1bbe6208927e21b78041f4a