perfetto-ui: Make track selection work with pinned tracks
You can now select pinned tracks just as you select regular tracks.

This CL also changes the logic for drawing the box selection,
previously we were figuring out the snapping logic based on where
the mouse was, separately from the logic that figures out which tracks
are in the region. This was unnecessary extra work since we can
instead use the selected tracks that have already been calculated &
find the min and max Y values to know where to draw the box.

Change-Id: I0509a0f39683e3a4cda506f4e5f2fb550bb433f5