Re-enable service worker
This is a revert of aosp/1223959. The bug seemed to be caused
by an overly large cache TTL on the AppEngine configuration
that caused integrity failures upon installation, because
intermediate Google caches were stale.
The right long-term fix would be using cache-busting and name
files after their hash (e.g. frontend_bundle-aa1f2c3.js) but
that is way more convoluted.
This CL removes the Crash dialog when the service worker
installation fails. That's too fragile (we later found that
an update causes the error on other tabs) and doesn't really
affect the functionality of the UI.

Bug: 148675312
Bug: 147015767
Change-Id: I9913144a5cfefe894c253aad838ee4c6386a7512