processor: Allow fallback to tid/pid_overrides for new-style TrackEvents
Importer code was previously not allowing pid/tid overrides for
TrackEvents with a "type" (as opposed to a legacy phase). However, the
client library will support pid/tid overrides for these types of events
(aosp/1216376) - at least temporarily -, because pid overrides cannot
be achieved with track_uuid in the presence of process-specific
randomness in the uuids. Thus we add support for these overrides in TP.

Note that we still disallow a fallback to the sequence's default
pid/tid for these types of events if track_uuid is set explicitly.

Bug: 130786981
Change-Id: Ifdad5895105ad6415decdcd54d4780c8ab9f0ab3