Client library: make code gc-sections friendly
The problem we want to solve is the following: on Android
some targets (e.g. libsurfaceflinger) rely on the system
backend for production code but then want the in-process
backend for tests. Both backends are big in terms of binary
size (~200 KB).
Instead of relying on dedicated build targets this code
arranges the build targets in a way that allows the linker to
dead-code-eliminate unused backends if the caller doesn't
require them.

Test: manual: the size of /system/lib64/libsurfaceflinger.so
      is == the size when completely removing the in-process
      target.
Bug: 148198993
Change-Id: Ie891a9b2e6ec38a3781318d1204a694899b01656