tp: add helper virtual table for computing duration from counter
This CL adds a virtual table to compute the duration from the counters
table directly without using window functions.

Using window functions makes this operation very expensive which is not
good for such an important table which we need to span join with other
tables.

A follow up CL will start using this table in the UI and metrics.

Change-Id: Ie8c111c953da5ad8190fc2b4cbe70be9cc32ae9a