perfetto-ui: Fix thread state duration
With most other tracks we can span join with the window table to get
just the visible slices. We send the id of the selected slice through
to the selection controller and then we can query for the real duration
etc. This doesn't work for thread state because there is no id that
allows for easy fetching. Instead we use the duration from the data
being displayed, this means that we need to have the full duration
of the slices at the beginning and end of the visible window - not
just chop them off. The ts constraints allow for this.

Also fix a canvas bug where if the line goes too far off either side
of the canvas (~+/-4m pixels) the line disappears.

Bug:140256335
Change-Id: I0bb5f57cddf1698ca2c046a151aaa81075bfe79c