traced_perf: record presence of ring buffer data loss in the trace
As mentioned in the comments, this wouldn't be particularly useful if we
wanted to give precise time bounds of "when" the loss happened (as the
loss is relative to a given ring buffer's contents). But I believe that
it's sufficient for now to distinguish that a given cpu had "some" loss
(since a typical trace should have no ring buffer loss).

Bug: 144281346
Change-Id: Iadd8d67a724caa45d6b032744534818b3f0f6cfa