processor: Parse StreamingProfilePacket timestamps during tokenization
Oystein discovered that resolving delta-timestamps during the parsing
stage is problematic, because a later ThreadDescriptor may reset the
reference value before its preceeding packets are parsed.

Simplest fix seems to be to move resolution of StreamingProfilePacket
reference timestamps into the tokenization phase. See also discussion
on aosp/1234246. Resolution of pid/tids has to remain in the parsing
phase, so we only insert into tables then.

This patch moves StreamingProfilePacket parsing into a new
ProfileModule to accomplish this. It also makes a few minor changes to
the way in which we register modules, to simplify test setup.

Bug: chromium:1046918
Change-Id: I34868e0ba9365bc050a87b654975f3089d33f341