tp: add support for dynamic tables and migrate flamegraph/counter dur
This CL introduces proper support for dynamic tables in SqliteDbTable
instead of the hacky thing we were doing before this point which was
subclassing and using parts of SqliteDbTable in arbitrary ways (e.g.
passing nullptr as the table and then overriding all methods, creating
a cursor but not using any of the rest of the code etc).

This prepares the way for describe_event in go/perfetto-analysis-design
which will be implemented as a dynamic table similar to the flamegraph
and counter-with-dur ones.

Change-Id: I8cff6554a2f54afa92dbf64c3eac7e6f70fbbdba