Integration test: don't force-disable ftrace on test runs
The integration test used to force-reset ftrace by writing
0 to tracefs/tracing_on before and after starting each test,
to isolate test failures.
Unfortunately this doesn't work when other production daemons
that require tracing (namely: iorap) are running in the system.
Doing so screws up the internal state of the ftrace muxer and
leads it into believing that some external entity is messing
with ftrace, causing it to bail out.

Bug: 149717312
Bug: 149802754
Test: atest CtsPerfettoTestCases
Change-Id: Ic7f737370b227f1b44f6aaae08a813326f72bfe1