perfetto-ui: Correct resolution upon loading
The resolution set by the controller (a guess) was not being overidden
with the correct resolution until much later when the user pans/zooms
etc. This CL corrects that and adds the correct lastUpdate when
the resolution is updated.

This bug was discovered while looking into b/149831017 but it's unclear
if it will solve it.

Change-Id: I080acb2ad800bfc72f79771c877ead92306b2133