test: log gtest onto logcat
This fixes the long-standing un-debuggability of tests running on the
Android test infrastructure. Because gtest are ran as host-side tests
the gtest log goes into another file (!= logcat) which is not time-synced
with logcat and makes impossible to reason about test failures.
This CL registers an extra gtest listener that injects a copy of the
markers into logcat.

Bug: now will be better logged
Test: manual
Change-Id: I56a7b284873ca1b59ae640df2820cb2ecf77777d