core: Add DCHECKs for process ID to TraceWriter (+1 in TraceBuffer)
While debugging startup tracing v2 crashes in Chrome, I discovered that
Chrome was enabling startup tracing before forking a child process that
ended up emitting a trace event. That's obviously broken because the
child process ended up writing to the same SMB as its parent process
(manifesting itself as bad chunk metadata in the service)...

This patch adds a DCHECK to TraceWriter to catch such a situation in
the future.

Also adds another DCHECK to TraceBuffer which turned out useful.

Change-Id: I0c7405f32816502371bb64942c059c3ab28d300b