traced_perf: enable SMB scraping + cts improvements
Scraping to help with timeouts on StopDataSource (as the producer
doesn't ack the stop until all of the internal buffers for that source
have been processed), which appears to be quite common on some test
targets (at least the gphone emu). Scraping shouldn't hurt outside
of tests either.

Fix an oversight in the cts test where it wasn't being skipped for older
kernels.

Otherwise, tweak timeouts and add a bit more logging on failures to see
whether cuttflefish is just slow, or there's an actual problem.

Change-Id: I862e82deee1f790afea942dacc767229e980e317
Bug: 149895879