ObservableEvents: use the enum defined in the proto
Non-funcional refactoring to ObservableEvents. Instead
of having two enums (one in the proto and one in
TracingService), just use the enum in the proto. This is
easier these days that we use the autogen proto->headers.
This avoid the need of manual plumbing for each new enum.
This will require a matching manual roll in Chromium
(http://crrev.com/c/2092531)

Test: updated existing tests
Bug: 150960749
Change-Id: I26e92d61ac286f022dba6f6e19c0d9ec76540759