perfetto-ui: Color summarized userspace slices
Summarized slices are colored based on the slices within them.
I considered changing these queries to be similar to the new thread
state ones, but because depth also has to be taken into account here
the new versions of the queries were too slow.
So only minimal query changes have been made.

One difference is that we used to specify on the controller side that
every summarized slice is the dur/width of a bucket, now the dur/width
is just the regular sum of the dur and slices are always drawn as at
least 1px on the frontend side. I think this looks more accurate from
the user perspective.

Bug:148446682
Change-Id: Ib6257201dda291a3f8fef47a091599a2f0fd233e