arbiter: Support aborting a startup tracing session.
It's possible that e.g. the tracing session associated with a startup
session is stopped before the producer successfully connected to the
tracing service. In this case, the producer's data sources will not be
started - Chrome detects this via a timeout. When this happens, we
should mark the startup target buffers as resolved to an invalid buffer
and clear away chunks from the SMB as best we can.

This patch adds a method to SharedMemoryArbiter to support aborting a
startup buffer reservation, which will associate the reservation with
an invalid target buffer, so that the tracing service will disregard
these chunks (but free them up) when they are committed.

Bug: 151203375,chromium:1060501
Change-Id: I83602947e0d33682df37cd7e46dd662defdc1d37