Check for build type, not debuggable, for extra guardrails test.
We were testing for ro.debuggable before, which is wrong in some weird
presubmit targets.

extra guardrails are rejected for fork mode, which is used on user
builds regardless of ro.debuggable. For release apps, the signal is
ignored based on ro.debuggable, so we keep using IsDebuggableBuild for
the condition in ReleaseApp(Runtime|Startup).

Bug: 152839251

Change-Id: Ia89339241b81d9f95f295c728220f51ed4356c9e