Introduce PERFETTO_COMPONENT_EXPORT macro
The PERFETTO_COMPONENT_EXPORT is meant for exporting symbols that are
defined in Perfetto's headers (e.g., perfetto::DataSource) but
instantiated by the user and therefore not exportable by Perfetto itself.
An example of where this is needed is a component build of v8:

- v8_perfetto.so builds Perfetto itself and uses the normal
  PERFETTO_EXPORT macro to export Perfetto's public symbols to the rest
  of V8.
- v8_libbase_without_compiler.so defines the track event categories used
  by V8 by instantiating perfetto::TrackEventCategoryRegistry. It can't
  use PERFETTO_EXPORT, because in this component that macro is defined
  to *import* symbols from v8_perfetto.so.
- v8.so wants to emit track events, so it needs to import the tracing
  categories from v8_libbase_without_compiler.so.

The solution is to set PERFETTO_COMPONENT_EXPORT=V8_EXPORT_PRIVATE in
v8_libbase_without_compiler.so, so that the tracing categories get
correctly exported to the rest of V8.

Ideally with this change we'd still let the compiler automatically
instantiate static template members for data sources, but there's a very
subtle pitfall here: if *any* part of the static member's definition
depends on a non-exported name, the compiler will use a weak symbol for
the member. This means you can easily end up having separate static
state in different modules of a component build, and this is very hard
to notice. This case is very easy to hit in a leaf node that is supposed
to import the static symbols from another .so -- here
PERFETTO_COMPONENT_EXPORT should also be set to make the implicitly
generated symbols public; otherwise the compiler will happy generate a
separate hidden copy.

To avoid having the user very carefully set PERFETTO_COMPONENT_EXPORT
correctly in all their data source definitions and usage, we'll
instead un-deprecate the PERFETTO_DEFINE_DATA_SOURCE_STATIC_MEMBERS
pattern, i.e., go back to explicit static data definition for data
sources. This means anyone using a custom data source needs these
definitions:

my_data_source.h:
  PERFETTO_DECLARE_DATA_SOURCE_STATIC_MEMBERS(MyDataSource);

my_data_source.cc:
  PERFETTO_DEFINE_DATA_SOURCE_STATIC_MEMBERS(MyDataSource);

Note that this isn't stricly a backwards incompatibility; the above
definitions are only needed in a component build setup.

This patch also adds some missing PERFETTO_EXPORT specifiers to various
symbols.

Bug: 148779455
Change-Id: Ic9673abe6bedae56b938adf364e209af181d9008