processor: Resolve descriptor track names during tokenization as well
I'm about to switch "EarlyJava" events in chrome to use descriptor
tracks for thread time. These events will have timestamps that precede
any track descriptors' timestamps. So the parser will not see the track
descriptor before the events.

In order to import the thread time from the extra_counter_values into
the thread-time column used by the JSON export, we need to know the
thread time track's name at the time of parsing the event. Thus, this
patch adds the track name to the track reservation and sets it initially
when resolving a descriptor track, too. The parser will still later
update the track names from descriptors in the sorted-by-ts order.

Change-Id: Ia8f5303aaa4c5c317537ff78e271d9199893d1cd