bazel: fix compile when precompiled headers are used
This CL fixes building the cpp proto targets when precompiled headers
are being used.

The build was failing because the generated headers are not standalone
when having a repeated field defined in another file.

Use textual_hdrs option to indicate to Bazel it should not try and
precompile these headers.

Also fix a small bug in the protogen rule where we returned all files
(including headers) as the default sourceset of the protogen rule.

Change-Id: I385b4bb8cb21565ef9ebd171aa0021d08821591a