Mark exhaustive time_in_state snapshots
In order to aggregate across frequencies, we need to know when we have a
complete view of time_in_state. The view is exhaustive after every
clearing of the incremental state and whenever a new thread is scanned
for the first time.

A marker saying that a snapshot is incremental (i.e. it is missing some
frequencies with non-zero ticks) is used instead of reporting each
frequency individually with 0 ticks to save space because at a given
time the majority of values are 0s.

Bug: 153092449
Test: perfetto_unittests
Change-Id: I9aecf9dbeb73af870b4779108d6582fe6c04dc84