Frame lifecycle MVP updates
We have decided to show the lifecycle of a frame as a view of ownership,
which makes it easy to understand for the users who are not
SurfaceFlinger-savvy. To do this, we try to create the ownership
slices(Phases) from the existing buffer events. This calls for a bit
more bookkeeping than usual.
Along with the updates, this CL also includes the following:
1) Move frame events parsing out of the graphics_event_parser
2) Rename the existing graphics_event_parser as gpu_event_parser as it
has only gpu events after the move
3) Code cleanup of the existing frame events parsing

Test: diff_test_trace_processor --trace-filter='graphics_frame*'
Bug: 155242423

Change-Id: I89201b1416f8ff37c03658dbd1ad9509e12a8453