ui/tp: fix race in trace-processor --httpd and improve performance
Fixes a race in the http_rpc_engine.ts queueing logic. The race
would happen if a new request is enqueued between the
resp.arrayBuffer() continuation. In that case the continuation would
re-fetch the already-outstanding request.

It also improves the performances of HTTP+RPC mode. Turns out that
adding cache: no-cache to the request prevents also caching of the
CORS pre-flight response (i.e. the OPTION request). This kills
pipelining introducing extra latencies for each query.

Bug: 151735611
Change-Id: Ia37854de36df0bd1d72bfdf76f6cf57aa631dc38