trace_processor_shell: don't quit on CTRL-C
The CTRL-C behavior is inconsistent:  if pressed while running
a query it interrupts the query but leaves the shell open.
If pressed when no query is running it quits.
This causes a lot of frustration when hittiing CTRL-C just
after a query has finished, especially on large queries.
This change makes it so CTRL-C doesn't quit ever.

Test: manual
Change-Id: I4fc6b0ac72754475b4b167c26abec48b3ed9c7a7