Detect main threads
The name of the main thread (tid = pid) comes from /proc/tid/comm. It is
typically a suffix of a package name. Mark such threads with as
main_thread=true in the thread table.

The time_in_state metric exports this marker to easily identify main
threads in the dashboard.

Bug: 157216280
Test: tools/diff_test_trace_processor.py
Change-Id: Ied8bf6bd6edbe49db0dc3ef3a09b2a9f14c01a36