tracing: Allow empty call to OnTraceData
This change makes it legal for a TracingBackend to call OnTraceData with
zero bytes to signal the end of trace data. This makes it easier to
implement a trace consumer backend in scenarios where trace data and
the EOF signal are delivered separately, i.e., when the backend knows
there's no more data to come, it has already sent all the data to the
consumer.

Bug: 158460267

Change-Id: Ibb918845a0d64d7890a954ad8308d4e8c84e712a