tp: rename _annotations suffix to _event
The _annotations name was there for historical reasons (because
go/perfetto-analysis originally called this feature annotations). When
we renamed annotations to refer to describing slices, we never renamed
the code referencing this feature correctly.

Make this change before more people start relying on this feature and
increase the difficulty of renaming.

Change-Id: Ia124b2116dab3cee22f890a9baaed172a1525fe4