Separate client management from hooks.
Currently, we include the client in the same library as the malloc
hooks. This is incorrect, as the malloc hooks should ship within the
com.android.runtime APEX, while the client should ship with the
platform (or wherever the heapprofd executable ships from).

This can then also be used for ART object allocation callstack
tracking.

Diff of HEAD~1:malloc_hooks.cc and client_ext.cc:
https://gist.github.com/segfaulthunter/05b7f0253a79c18b851d595a5d1df2a5

Test: get system_server profile.

Change-Id: If2b96e9a8c6baadcac8eea999653c6dd16977dca
Bug: 141241849